BF: KB now uses key object clock for collecting response timing
Instead of using clock created in the JS script, we now collect RT from
the key object in the event class. As the key object uses Monotonic clock
for reference time (where clock starts from the beginning of the script)
we need to account for this difference on every trial. First, a new
attribute is created in the KB object to hold the reference time. Then
we overwrite the clock method with an anonymous function, which
sets the baseTime attribute to that of the Monotonic clock ref time. This
is can then be called using the callOnFlip method, giving an accurate
ref time. The Key object returns time in seconds / 1000, so this is
accounted for when setting the kb rt time.