BF: Call nextEntry for routines when not in loop.
This fix will change how data is structured in the datafile. It saves
a new row of data for each routine not in a loop. This fix came about
whilst looking to fix another problem, where routine data was not saved
if no loops came after that routine (because nextEntry was not called).

Note, this current implementation will not work until PsychoJS
Experiment Handler attributes of `_trialsData` and 'thisCurrentTrial`
are changed from  to `entries` and `thisEntry`.